Re-introducing cfg::Instr. Marked extern funcs as not builtin so that they won't conflict with builtins needed to implement fpu instructions. Updated the address translation to have the scale be signed, and to implement the displacement as an add for positive numbers, and a subtract for negative numbers.

+++ remill/Arch/X86/Arch.cpp
@@ -333,7 +333,7 @@ static void DecodeMemory(Instruction *instr,
   op.addr.segment_reg = ReadReg(segment);
   op.addr.base_reg = ReadReg(base);
   op.addr.index_reg = ReadReg(index);
-  op.addr.scale = scale;
+  op.addr.scale = static_cast<int64_t>(scale);
   op.addr.displacement = disp;
 
   // Rename the base register to use `NEXT_PC` as the register name.
